Need to make checkpoint symlink for adjoint test relative
I guess that's the only confusing thing about the setup here: while we
specify the dependency as an absolute path from the project root, a
symlink is relative to *where it is*, so it's relative to the makefile
itself. That leads to a bit of a disconnect in how things are set up.