Fix up iv code for multiple internal variables
also add ci test (hopefully) with compressible viscoelastic
and elastic maxwell (1 internal variable) plus the same
compressible viscoelastic maxwell case modelled with 2 internal
variables with eta_1=eta_2 = eta_maxwell/2 and same for shear
modulus.

viscoelastic case as coarser grid than paper so that it is not
really slow on one core... convergence is 0.96 and 0.98 for 3 runs.