Fix LFS path in .gitattributes (#266)
I guess because we're not (currently) using this test, the fact that its
LFS checkout hasn't been working since we migrated the demo structure
has never come up! The object it references is still on the server, so
after fixing this we can `git lfs checkout` to get the file in place.