Update multi-material demos and tests (#270)
In this PR, I revisit all existing multi-material demos and tests,
improving them based on my experiences with the software since the test
suite was introduced. Notable changes:
- demos are now tested against expected values
- temperature field defined on a discontinuous space
- better mesh refinements throughout all demos and tests
- more meaningful simulation run times

The main goals are robustness and reliability of the testing framework
and improved clarity of demos.

New tests and demos will be introduced in a follow-up PR.

---------

Co-authored-by: Rhodri Davies <rhodri.davies@anu.edu.au>