Fix weak velocity bcs for compressible
The weak "un" (Nitsche bc) boundary condition was incorrect for
compressible as one of the additional div(u) terms was accidentally
multiplied by 2. Weak "u" bcs were also incorrect, but these are
currently not actually accesible as we always turn bc type "u" into
strong `DiricletBC`s.

This was noticed due to asymmetry of the velocity matrix (which should
still be symmetric for compressible). Thus adding a test that
tests for symmetry for the various cases we support.

Co-authored-by: dsroberts <115434035+dsroberts@users.noreply.github.com>