Minor maintenance issues associated with level-set code (#320)
In the thermochemical buoyancy demo, the value of `side` given to
`material_entrainment` is not the expected one: it should match the
level-set value of the dense material being entrained, which is 1.0.
Correcting this value yields an entrainment value of 0.0, which matches
the fact that the demo lacks mesh resolution to resolve tendril
entrainment.

Following the release of Firedrake 2025.10, Fanny and I noticed the
following error when running our subduction script:
```python
Traceback (most recent call last):
  File "/scratch/xd2/fg9926/tmp/SUBDUCTION/test2bis/subduction.py", line 638, in <module>
    level_set_solver = LevelSetSolver(psi, adv_kwargs=adv_kwargs, reini_kwargs=reini_kwargs)
  File "/scratch/xd2/fg9926/tmp/g-adopt/gadopt/level_set_tools.py", line 491, in __init__
    self.set_up_solvers()
  File "/scratch/xd2/fg9926/tmp/g-adopt/gadopt/level_set_tools.py", line 563, in set_up_solvers
    self.adv_solver = GenericTransportSolver(
  File "/scratch/xd2/fg9926/tmp/g-adopt/gadopt/transport_solver.py", line 311, in __init__
    super().__init__(solution, delta_t, timestepper, **kwargs)
  File "/scratch/xd2/fg9926/tmp/g-adopt/gadopt/transport_solver.py", line 137, in __init__
    self.setup_solver()
  File "/scratch/xd2/fg9926/tmp/g-adopt/gadopt/transport_solver.py", line 237, in setup_solver
    self.ts = self.timestepper(
  File "/scratch/xd2/fg9926/tmp/g-adopt/gadopt/time_stepper.py", line 144, in __init__
    super(ERKGeneric, self).__init__(
  File "/scratch/xd2/fg9926/tmp/g-adopt/gadopt/time_stepper.py", line 75, in __init__
    self.dt = float(dt)
TypeError: Division.__float__ returned non-float (type NotImplementedType)
```
I identified the issue as coming from the time step provided to
`GenericTransportSolver`, which involves a division. Ensuring the time
step was given as a float solved the issue, but I could not quite
understand why the error arose in the first place. In other words, I did
not manage to reproduce it outside the script.

Other minor changes are of formatting and typographic nature.