Add multi-material test from Woidt (1978) (#318)
The setup for this test is described in [Woidt
(1978)](https://www.sciencedirect.com/science/article/pii/0040195178901439)
(see Figure 8). It is a compositional buoyancy test involving two
materials with distinct densities and viscosities. The setup is
dimensional, and the numerical domain has a high-aspect ratio.
Initialisation of the level set involves a custom curve callable. To
reproduce the results presented in Woidt (1978), the averaging scheme at
the interface for density needs to be set as `arithmetic`. I encountered
this test case when I was looking for benchmarks with analytical
solutions. The analytical solution mentioned in Woidt (1978) is only
valid for small deformation (i.e. at about 10 Myr here), unfortunately,
so it is not featured in my implementation. Nevertheless, I found the
test case interesting and singular enough to propose including it in
G-ADOPT.


https://github.com/user-attachments/assets/aefd21b4-1394-44b1-8ce6-6e0831590f93