Add multi-material demo with free surface based on Schmeling et al. (2008) (#331)
This PR adds another multi-material demo that focuses on the interaction
between compositional buoyancy and a free surface under the Boussinesq
approximation. Although the evolution is sensible and reproduces
published results, I am concerned by the fact that, in order to
reproduce these results, I had to specify a material interface at the
top boundary of the lithosphere. If I do not do that, then subduction
becomes sluggish. My interpretation is that imposing a material
interface at the top boundary decreases the viscosity there because the
mantle has a lower viscosity than the lithosphere, and this viscosity
decrease is needed to enable subduction.

Animation of level-set and free-surface evolutions:


https://github.com/user-attachments/assets/44b610a5-ac5a-43d8-a819-078b6d274aac