Add mesh adaptive version of the base case demo. (#334)
Exact same set up as the mantle convection base case, except at Ra=1e5
to make it a little less boring, but now with mesh adaptivity.

Someone probably needs to take a good look at the pyvista bit, as it's
the first time I'm using it - so I don't know what I'm doing (movie
doesn't appear in the notebook - is it meant to? - just written to .gif)

This is currently based of release, but I don't know if that's
appropriate. This also depends on what we need to do to ensure
mmg+animate are available (@angus-g ?). How are the demos run at the
workshop, is this co-lab? Do we know how to use our own petsc+firedrake
with mmg in there?

I'm not a 100% sure what the status of animate+"firedrake main" is. I
finally switched over its CI to firedrake main, and all basic API
changes have been addressed, but there's still some tests failing. This
unfortunately also means animate no longer works with "firedrake
release", and we don't have an official "animate release" yet but I can
provide a branch. Let me know which one of the two you want me to
prioritise (firedrake main or release) for the workshop.

---------

Co-authored-by: Angus Gibson <angus@agibson.me>