Warn user when `constant_jacobian=True` (#333)
It has happened a few times that simulations are run with
`constant_jacobian=True` given to `StokesSolver` when the Jacobian was
not actually constant (e.g. when the viscosity varies in time). This PR
proposes to raise a warning when this argument is supplied, such that a
user can ascertain whether or not they meant to include it.