Streamline `time_stepper.py` (#345)
In this follow-up PR to the Irksome integration, I propose some
simplifications to the way classes are defined in the `time_stepper.py`
module. I also propose some changes to comments, docstrings, and the
unit test script.

Summary of main changes:
- All time-stepper schemes are now directly exposed by G-ADOPT via
`__all__` (allowing `import *` in `__init__.py`, which we should
definitely use for all modules), and their names have been simplified
(e.g. `CrankNicolson` instead of `CrankNicolsonRK` or `RadauIIA` instead
of `IrksomeRadauIIA`).
- `equation` can now handle a single `Equation` instance, a list of such
instances (in which case `solution` should live in a mixed space), or a
`Form`.
- `RKGeneric` now accepts the `tableau_parameter` argument, which allows
customising the initialisation of certain Irksome tableaux.
- `StaticButcherTableauIntegrator` has been removed, as it was not
needed.
- `AbstractRKScheme` now uses `__init_subclass__` to perform
initialisation of children schemes independently of provided arguments
(which are passed to `RKGeneric`).
- `eSSPRK` schemes now share a common parent class, simplifying scheme
definition (which was quite verbose before).
- All Irksome tableaux now have streamlined definitions without the need
of an explicit `__init__`.

---------

Co-authored-by: Dale Roberts <dale.roberts1@anu.edu.au>